feat: handle errors

+++ src/cli/files/deleteFile.js
@@ -2,12 +2,16 @@ import path from "path";
 import fs from "fs/promises";
 
 const deleteFile = async (args) => {
-  try {
-    const pathToFile = path.join(process.cwd(), args[0]);
-    await fs.access(pathToFile, fs.constants.F_OK);
-    await fs.unlink(pathToFile);
-  } catch {
-    console.error("Operation failed");
+  if (!args.length) {
+    console.error("Invalid input");
+  } else {
+    try {
+      const pathToFile = path.join(process.cwd(), args[0]);
+      await fs.access(pathToFile, fs.constants.F_OK);
+      await fs.unlink(pathToFile);
+    } catch {
+      console.error("Operation failed");
+    }
   }
 };
