+++ src/cli/files/move.js
@@ -2,11 +2,15 @@ import fs from "fs";
 import { copy } from "./copy.js";
 
 const move = async (args) => {
-  try {
-    const originalFilePath = await copy(args);
-    await fs.promises.unlink(originalFilePath);
-  } catch {
-    console.error("Operation failed");
+  if (args.length < 2) {
+    console.error("Invalid input");
+  } else {
+    try {
+      const originalFilePath = await copy(args);
+      await fs.promises.unlink(originalFilePath);
+    } catch {
+      console.error("Operation failed");
+    }
   }
 };
